refactor(ui): структурирован код экрана статей по отдельной директории
компоненты Articles Screen перенесены в Views/ArticlesScreen и разделены по ролям: state, presentation models, derived view state, actions и day sections builder
row и bulk actions вынесены из ArticleListView в ArticlesScreenActions, а часть presentation logic выделена из ArticlesScreenState в отдельные типы
такая структура уменьшает размер отдельных файлов и делает границы между screen state, screen behavior и UI composition более явными
