+++ RSSReader/Views/ArticlesScreen/ArticleListPreviewData.swift
